refactor(chat): simplify parseClarificationBlock filter; move test to __tests__/chat/
