+++ packages/backend/src/routes/api/chat/parse-clarification-block.ts
@@ -55,6 +55,5 @@ export function parseClarificationBlock(
     questions.push(current)
   }
 
-  const valid = questions.filter((q) => q.question.length > 0)
-  return valid.length > 0 ? valid : null
+  return questions.length > 0 ? questions : null
 }
